Always use LINK refs to support delete
Store plaintext data size in Ref and allow ChunkSize to be requested in
Header

This may allow for better sharing from some kinds of data, truncates to
a MaxChunkSize based on GRPC messages size.

Storing size in individual refs allosw for chunk-wise random access down
the line

+++ chunking.go
@@ -6,7 +6,7 @@ import (
 	"io"
 )
 
-func CopyChunked(dest func(chunk []byte) error, src func() ([]byte, error), chunkSize int) error {
+func CopyChunked(dest func(chunk []byte) error, src func() ([]byte, error), chunkSize int64) error {
 	// The internal Buffer will ensure we write in chunks
 	_, err := io.CopyBuffer(NewPusher(dest), NewPuller(src), make([]byte, chunkSize))
 	if err != io.EOF {
