+++ cmd/hoarctl/store.go
@@ -71,7 +71,7 @@ func (client *Client) Insert(cmd *cli.Cmd) {
 	chunk := addIntOpt(cmd, "chunk", chunkOpt, chunkSize)
 
 	cmd.Action = func() {
-		validateChunkSize(*chunk)
+		validateChunkSize(int64(*chunk))
 
 		// If given address use it
 		push, err := client.storage.Push(context.Background())
@@ -81,7 +81,7 @@ func (client *Client) Insert(cmd *cli.Cmd) {
 
 		var addresses []*api.Address
 
-		err = hoard.NewStreamer().WithChunkSize(*chunk).
+		err = hoard.NewStreamer().WithChunkSize(int64(*chunk)).
 			WithInput(os.Stdin).
 			WithSend(func(data []byte) error {
 				return push.Send(&api.Ciphertext{EncryptedData: data})
@@ -111,7 +111,7 @@ func (client *Client) Put(cmd *cli.Cmd) {
 	chunk := addIntOpt(cmd, "chunk", chunkOpt, chunkSize)
 
 	cmd.Action = func() {
-		validateChunkSize(*chunk)
+		validateChunkSize(int64(*chunk))
 
 		put, err := client.cleartext.Put(context.Background())
 		if err != nil {
@@ -124,7 +124,7 @@ func (client *Client) Put(cmd *cli.Cmd) {
 		}
 
 		refs := reference.Refs{}
-		err = hoard.NewStreamer().WithChunkSize(*chunk).
+		err = hoard.NewStreamer().WithChunkSize(int64(*chunk)).
 			WithInput(os.Stdin).
 			WithSend(func(data []byte) error {
 				return put.Send(&api.Plaintext{Body: data})
